MLK-13721 usb: chipidea: otg: add timeout for wait usb device disconnect
While quickly switch the OTG port between host and gadget by ID cable plug
and unplug with a usb net device, the disconnect event may be lost, but
we have a wait-for-ever on usb device disconnection before stop host role
when switch from host to gadget, so this wait will not finish and the irq
not be enabled in this case, in the end this port can't be used at all.

This patch adds a wait timeout(6s~9s) to let a normal disconnect event
finished by host driver, if timeout, we just start to stop host anyway.

Acked-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>