MLK-18042-02 arm: dts: update the backlight brightness on imx7ulp evk
On the i.MX7ULP EVK Rev.B baord, the backlight brigntness driver circuit
is updated. A RC filter is added on the MP3301's EN pin. So the PWM's
frequency should be change to 20KHZ. for EN pin, A DC voltage from 0.7V
to 1.4V can control the LED current from 0% to 100%. the backlight
brightness level also need to be updated.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Reviewed-by: Anson Huang <Anson.Huang@nxp.com>
Signed-off-by: Arulpandiyan Vadivel <arulpandiyan_vadivel@mentor.com>
Signed-off-by: Srikanth Krishnakar <Srikanth_Krishnakar@mentor.com>