MLK-18628-2 ARM: dts: imx7d-sdb: Restore &pwm1 node
This was removed by upstream commit 5eaeaccdaedb
("ARM: dts: imx7d-sdb: Pass 'enable-gpios' and 'power-supply' properties")

That commit claims that GPIO1_IO01 is not a PWM because it's connected
to a pin labelled "PWREN" of J14 connector. However that pin does behave
as a pwm despite the name.

The exact same pattern happens with imx6ul-evk which has an LCD8000-43T
display on J901.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Fabio Estevam <fabio.estevam@nxp.com>
Signed-off-by: Arulpandiyan Vadivel <arulpandiyan_vadivel@mentor.com>
Signed-off-by: Srikanth Krishnakar <Srikanth_Krishnakar@mentor.com>