MLK-21397 drm/imx: dpu: kms: Put possible pipe states properly
We should put possible pipe states after we've finished assigning
resources for all pipes.  This makes sure the pipe states which
should be put are all eventually put.  Without this patch, disabled
crtc(s) or enabled crtc(s) without any plane upon will be accidentally
left in the atomic state when commit operations only happen for other
pipe(s), because we skip the put operations when we see the relevant
active plane number is zero.  It's good to put pipe states for those
crtcs, otherwise, we'll depend on atomic helper core behavior to
handle the crtcs.

Signed-off-by: Liu Ying <victor.liu@nxp.com>