MLK-21989: VPU Decoder: avoid crash after vpu decoder hang
If vpu is hang, driver won't really release the instance when v4l2_release,
but driver will release queue and buffer,
the sysfs device instance*_buffer isn't removed.
when we try to access the instance*_buffer in such case,
kernel crash will occur.
Driver release queue and buffer when we really release the instance,
and the crash can be avoided

Signed-off-by: ming_qian <ming.qian@nxp.com>