MLK-21819: drm/bridge: nwl-dsi: Use sync events instead of sync pulse
For some reason, when sending HS DSI commands while also sending pixel
data, the display timing becomes messed up. Didn't find the reason for
which this is happening, but it seems that setting the video mode for
DPI to Non-Burst with Sync Events instead of Non-Burst with Sync Pulses
allows the insertion of HS DSI commands to be send without problems.
So, use Sync Pulses only if requested by mode_flags and use Sync Events
as default.
Also:
- added defines for this video mode, in order to explain those magic
numbers used
- removed write to ENABLE_MULT_PKTS register twice

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>